pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kubesecret"
version = "1"
description = "CLI tool for kubernetes secrets"
readme = "README.md"
requires-python = ">=3.6"
dependencies = [
    "click",
    "rich",
    "pyyaml",
]
authors = [
    {name = "Ernesto Vargas"}
]
license = {text = "Apache License, Version 2.0"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: System :: Systems Administration",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/netoxico/kubesecret"

[project.scripts]
kubesecret = "kubesecret.cli:cli"
